requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0
motor==3.3.2
requests==2.31.0
jinja2==3.1.3
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
